refactor: activities of the day into multiple components
